+++ lib/macros/buttonPadding.dart
@@ -6,8 +6,6 @@ abstract class ButtonPaddingMacro {
   _initialize(Function callback) {
     if (buttonPadding == null) buttonPadding = new NikuEdgeInsets.all(20);
 
-    print(buttonPadding?.value);
-
     callback();
   }